update test cases for Math.Pow() vs Math.Exp()

+++ CommonProblems/CommonProblems/Maths/EfficientPower.cs
@@ -3,19 +3,20 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-
+using System.Numerics;
 namespace CommonProblems.Maths
 {
 	public class EfficientPower
 	{
-		public Int64 NormalPowerMethod(double x, int toPower)
+		public double NormalPowerMethod(double x, double toPower)
 		{
-			return (Int64)Math.Pow(x, toPower);
+			return Math.Pow(x, toPower);
+			
 		}
 
-		public Int64 EfficientPowerMethod(double x, int toPower)
+		public double EfficientPowerMethod(double x, double toPower)
 		{
-			return (Int64)Math.Exp(Math.Log(x) * toPower);
+			return Math.Exp(Math.Log(x) * toPower);
 		}
 	}
 }
